feat(cart): finished consolidating cart and order reducers
